agents/history_agent/requirements.txt
google-adk>=1.2.0
google-cloud-aiplatform[agent_engines,adk]>=1.88.0
google-auth>=2.20.0
httpx>=0.27.0
pyyaml>=6.0
